Refacto to use dto file to unify models
